rag-chatbot/requirements.txt
fastapi
uvicorn
pymongo
faiss-cpu
langchain
langchain-huggingface
huggingface-hub
sentence-transformers
python-dotenv
